+++ packages/data-ferret/src/lib/registerIterableClass/registerIterableClass.spec.ts
@@ -12,13 +12,15 @@ describe('registerIterableClass', () => {
       Map,
       map => Array.from(map.keys()) as string[],
       (map, key) => map.get(key),
-      (map, value, key) => map.set(key, value)
+      (map, value, key) => map.set(key, value),
+      (map, key) => map.delete(key)
     )
     registerIterableClass<Set<unknown>>(
       Set,
       set => Array.from(set.keys()) as string[],
       (_, key) => key,
-      (set, value) => set.add(value)
+      (set, value) => set.add(value),
+      (set, key) => set.delete(key)
     )
     expect(registeredIterableClasses.length).toEqual(4)
     const registeredIterableClassRefs = registeredIterableClasses.map(e => e.classRef)
@@ -34,13 +36,15 @@ describe('registerIterableClass', () => {
       map => Array.from(map.keys()) as string[],
       (map, key) => map.get(key),
       (map, value, key) => map.set(key, value),
+      (map, key) => map.delete(key),
       () => new Map()
     )
     registerIterableClass<Map<unknown, unknown>>(
       Map,
       map => Array.from(map.keys()) as string[],
       (map, key) => map.get(key),
-      (map, value, key) => map.set(key, value)
+      (map, value, key) => map.set(key, value),
+      (map, key) => map.delete(key)
     )
     expect(registeredIterableClasses.length).toEqual(3)
     const registeredIterableClassRefs = registeredIterableClasses.map(e => e.classRef)
@@ -54,7 +58,8 @@ describe('registerIterableClass', () => {
       Set,
       set => Array.from(set.keys()) as string[],
       (_, key) => key,
-      (set, value) => set.add(value)
+      (set, value) => set.add(value),
+      (set, key) => set.delete(key)
     )
     expect(registeredClasses).toContain(Set)
   })