+++ packages/data-ferret/src/lib/removeKey/index.ts
@@ -1,0 +1,1 @@
+export { removeKey } from './removeKey'